orchestration/pyproject.toml
[project]
name = "dakota-orchestration"
version = "0.1.0"
description = "Prefect orchestration for Dakota technical assessment pipeline"
requires-python = ">=3.11"
dependencies = [
  "prefect>=3.0.0",
  "psycopg2-binary>=2.9.9",
  "dbt-core>=1.11.0",
  "dbt-postgres>=1.10.0",
  "matplotlib>=3.9.0",
  "reportlab>=4.2.0",
]

[tool.uv]
dev-dependencies = []
